+++ API/index.js
@@ -3,7 +3,7 @@ const { conn } = require('./src/db.js');
 const fillTables = require('./src/helpers/fillTables.js');
 const port = process.env.PORT || 3001;
 
-conn.sync({ force: true })
+conn.sync({ force: false })
   .then(() => fillTables())
   .then(() =>
     app.listen(port, () => console.log(`Listening on port ${port}!`))
